feat: Menu가격을 검증하는 MenuPrice도메인추가

+++ src/test/java/kitchenpos/product/domain/ProductPriceTest.java
@@ -4,6 +4,8 @@ package kitchenpos.product.domain;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
+
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 public class ProductPriceTest {
@@ -18,7 +20,7 @@ public class ProductPriceTest {
     @DisplayName("입력하는 돈이 0보다작으면 예외발생")
     @Test
     public void throwsExceptionWhenNetativeAmount(){
-        assertThatThrownBy(() -> ProductPrice.of(-1000l))
+        assertThatThrownBy(() -> ProductPrice.of(BigDecimal.valueOf(-1000)))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 }
